pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-synth-miner"
version = "0.1.0"
description = "Hydra-driven research framework for Synth hybrid mining"
readme = "README.md"
authors = [{ name = "Open Synth Miner Contributors" }]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    # Core frameworks
    "hydra-core>=1.3.2",
    "hydra-zen>=0.6.0",
    "omegaconf>=2.3.0",
    # ML stack
    "torch>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.2.0",
    "properscoring>=0.1",
    "pyarrow>=15.0.0",
    "PyWavelets>=1.5.0",
    "scikit-learn>=1.3.0",
    # Experiment tracking and hubs
    "wandb==0.25.0",
    "huggingface_hub>=0.22.0",
    # Utilities
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
sde = ["torchsde>=0.2.6"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "coverage-badge>=1.1.0",
    "torchsde>=0.2.6",
]

[tool.setuptools]
include-package-data = true
py-modules = ["main"]
packages = [
    "osa",
    "osa.data",
    "osa.models",
    "osa.models.components",
    "osa.models.components.attention",
    "osa.models.components.patches",
    "osa.research",
    "osa.tracking",
    "open_synth_miner",
]

[tool.setuptools.package-dir]
osa = "src"
"osa.data" = "src/data"
"osa.models" = "src/models"
"osa.models.components" = "src/models/components"
"osa.models.components.attention" = "src/models/components/attention"
"osa.models.components.patches" = "src/models/components/patches"
"osa.research" = "src/research"
"osa.tracking" = "src/tracking"

[tool.setuptools.package-data]
"*" = [
    "configs/*.yaml",
    "configs/**/*.yaml",
    "configs/**/*.json",
    "configs/**/*.txt",
]

[project.scripts]
open-synth-miner = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
